refactor(assessment-statistics): abstract translations
abstract out the translations of some assessment stsatistics files
add translations and localizations for the associated translations
